+++ Dendrite.cpp
@@ -2,7 +2,7 @@
 #include <stdexcept>
 
 
-Dendrite::Dendrite(std::shared_ptr<OutputNeuron> outputNeuron, std::shared_ptr<Axon> axon, const double multiplier): multiplier(multiplier){
+Dendrite::Dendrite(std::shared_ptr<OutputNeuron> outputNeuron, std::shared_ptr<Axon> axon, const long double multiplier): multiplier(multiplier){
 	if(outputNeuron == nullptr || axon == nullptr)
 		throw std::runtime_error("nullptr is passed");
 
@@ -10,12 +10,12 @@ Dendrite::Dendrite(std::shared_ptr<OutputNeuron> outputNeuron, std::shared_ptr<A
 	this->axon = axon;
 }
 
-void Dendrite::resendSignal(const double value){
+void Dendrite::resendSignal(const long double value){
 	this->outputNeuron->receiveSignal(value * multiplier);
 }
 
-void Dendrite::resendError(const double error, const double teachingSpeed){
-	double diff = error * teachingSpeed * this->axon->getInputNeuron()->calcSignal();
+void Dendrite::resendError(const long double error, const long double teachingSpeed){
+	long double diff = error * teachingSpeed * this->axon->getInputNeuron()->calcSignal();
 	this->axon->getInputNeuron()->getOffsetCorrection(error * this->multiplier);
 	this->multiplier += diff;
