+++ Dendrite.hpp
@@ -13,13 +13,13 @@ class Dendrite{
 protected:
 	std::shared_ptr<OutputNeuron> outputNeuron;
 	std::shared_ptr<Axon> axon;
-	double multiplier;
+	long double multiplier;
 
 
 public:
-	Dendrite(std::shared_ptr<OutputNeuron> outputNeuron, std::shared_ptr<Axon> axon, const double multiplier);
-	void resendSignal(const double value);
-	void resendError(const double error, const double teachingSpeed);
+	Dendrite(std::shared_ptr<OutputNeuron> outputNeuron, std::shared_ptr<Axon> axon, const long double multiplier);
+	void resendSignal(const long double value);
+	void resendError(const long double error, const long double teachingSpeed);
 
 	bool compareOutputs(std::shared_ptr<Dendrite> dendrite) const;
