+++ SuperNeuron.cpp
@@ -1,5 +1,5 @@
 #include "SuperNeuron.hpp"
 
-SuperNeuron::SuperNeuron(const double weight): Neuron(weight), InputNeuron(weight), OutputNeuron(weight){
+SuperNeuron::SuperNeuron(const long double weight): Neuron(weight), InputNeuron(weight), OutputNeuron(weight){
 }
